Refactors abac_tags var from map(string) to map(list(string))
Before we tried using a map(string) that would indicate both the tags to use when tagging as the valid ones to insert into the IAM Polices. Which was not the right approach as the tags used for tagging is a map(string) and the condition tags is a map(list(string)), because we have a list of accepted values for each key.